+++ src/test/clojure/clojure/tools/analyzer/passes_test.clj
@@ -3,12 +3,13 @@
   (:require [clojure.tools.analyzer.ast :refer :all]
             [clojure.test :refer [deftest is]]
             [clojure.set :as set]
-            [clojure.tools.analyzer.core-test :refer [ast e]]
+            [clojure.tools.analyzer.core-test :refer [ast e e1]]
             [clojure.tools.analyzer.passes.add-binding-atom :refer [add-binding-atom]]
             [clojure.tools.analyzer.passes.source-info :refer [source-info]]
             [clojure.tools.analyzer.passes.uniquify :refer [uniquify-locals]]
             [clojure.tools.analyzer.passes.constant-lifter :refer [constant-lift]]
-            [clojure.tools.analyzer.passes.emit-form :refer [emit-form emit-hygienic-form]]))
+            [clojure.tools.analyzer.passes.emit-form :refer [emit-form emit-hygienic-form]]
+            [clojure.tools.analyzer.env :refer [with-env]]))
 
 (deftest passes-utils-test
   (let [ast {:foo [{:a 1} {:a 2}] :bar [{:a 3}] :children [:foo :bar]}]
@@ -35,15 +36,16 @@
     (is (= [{:a 1} {:a 2} {:a 3}] (children ast)))))
 
 (deftest add-binding-atom-test
-  (let [the-ast (add-binding-atom (ast (let [a 1] a)))]
+  (let [the-ast (prewalk (ast (let [a 1] a))
+                         (partial add-binding-atom (atom {})))]
     (swap! (-> the-ast :bindings first :atom) assoc :a 1)
     (is (= 1 (-> the-ast :body :ret :atom deref :a)))))
 
 (deftest source-info-test
   (is (= 1 (-> {:form ^{:line 1} [1]} source-info :env :line)))
-  (is (= 1 (-> {:form ^{:column 1} [1]} source-info :env :column)))
-  (is (= 1 (-> {:form ^{:end-line 1} [1]} source-info :env :end-line)))
-  (is (= 1 (-> {:form ^{:end-column 1} [1]} source-info :env :end-column))))
+  (is (= 1 (-> {:form ^{:column 1 :line 1} [1]} source-info :env :column)))
+  (is (= 1 (-> {:form ^{:end-line 1 :line 1} [1]} source-info :env :end-line)))
+  (is (= 1 (-> {:form ^{:end-column 1 :line 1} [1]} source-info :env :end-column))))
 
 (deftest constant-lift-test
   (is (= :const (-> (ast {:a {:b :c}}) (postwalk constant-lift) :op)))
@@ -52,8 +54,9 @@
                   (postwalk constant-lift) :op))))
 
 (deftest uniquify-test
-  (let [the-ast (uniquify-locals (ast (let [x 1 y x x (let [x x] x)]
-                                        (fn [y] x))))]
+  (let [the-ast (with-env e1
+                  (uniquify-locals (ast (let [x 1 y x x (let [x x] x)]
+                                          (fn [y] x)))))]
     (is (= 'x__#2 (-> the-ast :body :ret :methods first :body :ret :name)))
     (is (= 'y__#1 (-> the-ast :body :ret :methods first :params first :name)))
     (is (apply not= (->> the-ast :bindings (mapv :name))))))
@@ -91,18 +94,20 @@
          (emit-form (ast (try (throw 1) (catch e t b) (finally 2)))))))
 
 (deftest emit-hygienic-form-test
-  (is (= '(let* [a__#0 1 a__#1 a__#0] a__#1)
-         (emit-hygienic-form (uniquify-locals (ast (let [a 1 a a] a))))))
-  (is (= '(let* [x__#0 1] (fn* ([x__#1] x__#1)))
-         (emit-hygienic-form (uniquify-locals (ast (let [x 1] (fn [x] x)))))))
-  (is (= '(fn* x__#0 ([x__#1] x__#1))
-         (emit-hygienic-form (uniquify-locals (ast (fn x [x] x)))))))
+  (with-env e1
+    (is (= '(let* [a__#0 1 a__#1 a__#0] a__#1)
+           (emit-hygienic-form (uniquify-locals (ast (let [a 1 a a] a))))))
+    (is (= '(let* [x__#0 1] (fn* ([x__#1] x__#1)))
+           (emit-hygienic-form (uniquify-locals (ast (let [x 1] (fn [x] x)))))))
+    (is (= '(fn* x__#0 ([x__#1] x__#1))
+           (emit-hygienic-form (uniquify-locals (ast (fn x [x] x))))))))
 
 (deftest deeply-nested-uniquify
   (is (= '(fn* ([x__#0 y__#0 z__#0]
                   (let* [foo__#0 (fn* ([y__#1 z__#1] [y__#1 z__#1]))]
                         (foo__#0 x__#0 y__#0))))
-         (emit-hygienic-form (uniquify-locals (ast (fn [x y z]
-                                                     (let [foo (fn [y z]
-                                                                 [y z])]
-                                                       (foo x y)))))))))
+         (with-env e1
+          (emit-hygienic-form (uniquify-locals (ast (fn [x y z]
+                                                      (let [foo (fn [y z]
+                                                                  [y z])]
+                                                        (foo x y))))))))))